+++ cyberdrop_dl/database/schema.py
@@ -32,7 +32,7 @@ class Version(NamedTuple):
         return ".".join(map(str, self))
 
 
-CURRENT_VERSION = Version(10, 1, 0)
+CURRENT_VERSION = Version(10, 2, 1)
 REQUIRED_VERSION = Version(9, 15, 0)
 
 logger = logging.getLogger(__name__)