+++ cyberdrop_dl/utils/dataclass.py
@@ -7,7 +7,7 @@ from cyberdrop_dl.constants import MISSING
 from cyberdrop_dl.utils import fast_cache
 
 if TYPE_CHECKING:
-    from collections.abc import Mapping
+    from collections.abc import Iterator, Mapping
 
 
 class _DataClass(Protocol):
@@ -18,12 +18,12 @@ _FIELDS_CACHE: dict[type, tuple[str, ...]] = {}
 
 
 @fast_cache
-def _fields(cls: type[_DataClass]) -> tuple[str, ...]:
+def _fields_names(cls: type[_DataClass]) -> tuple[str, ...]:
     return tuple(f.name for f in dataclasses.fields(cls) if f.init)
 
 
 def filter_data[DataClassT: _DataClass](cls: type[DataClassT], data: Mapping[str, Any], /) -> dict[str, Any]:
-    return {name: value for name in _fields(cls) if (value := data.get(name, MISSING)) is not MISSING}
+    return {name: value for name in _fields_names(cls) if (value := data.get(name, MISSING)) is not MISSING}
 
 
 def deserialize[DataClassT: _DataClass](
@@ -37,6 +37,10 @@ def deserialize[DataClassT: _DataClass](
 
 
 class DictDataclass(_DataClass, Protocol):
+    def __iter__(self) -> Iterator[tuple[str, Any]]:
+        for field_name in _fields_names(type(self)):
+            yield field_name, getattr(self, field_name)
+
     filter_dict = classmethod(filter_data)  # pyright: ignore[reportUnannotatedClassAttribute]
 
     @classmethod